# speed up track_pixel_map