This PR speeds up the computation of `track_pixel_map` by reusing the index arrays that were used to create `pixel_index_map` earlier.

This improves the runtime of the batched track simulation loop (excluding the first iteration) by ~ 1.5 s / 1.2 s = 1.25x for ntracks=2000.

```
$ python cli/simulate_pixels.py --input_filename examples/edepsim_1M.h5 --output_filename=test.h5 --n_tracks=2000 --pixel_layout=larndsim/pixel_layouts/layout-singlecube.yaml --detector_properties=larndsim/detector_properties/singlecube.yaml

  _                      _            _
 | |                    | |          (_)
 | | __ _ _ __ _ __   __| |______ ___ _ _ __ ___
 | |/ _` | '__| '_ \ / _` |______/ __| | '_ ` _ \
 | | (_| | |  | | | | (_| |      \__ \ | | | | | |
 |_|\__,_|_|  |_| |_|\__,_|      |___/_|_| |_| |_|


**************************
LOADING SETTINGS AND INPUT
**************************
Pixel layout file: larndsim/pixel_layouts/layout-singlecube.yaml
Detector propeties file: larndsim/detector_properties/singlecube.yaml
edep-sim input file: examples/edepsim_1M.h5
*******************
STARTING SIMULATION
*******************
Quenching electrons... 0.90 s
Drifting electrons... 0.51 s
Simulating pixels...: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:07<00:00,  1.26it/s]
- total time: 7.95 s
- excluding first iteration: 1.19 s
Writing to HDF5...: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20096/20096 [00:00<00:00, 22749.55it/s]
Output saved in: test.h5
run_simulation elapsed time: 18.34 s
```

Note that this PR is based on #4