@dmargala 2021-03-03
I've corrected this PR by reverting the calculation of track_pixel_map and then converting detsim.get_track_pixel_map to a cuda kernel. The output is now identical to the output from the base branch as measured by `h5diff`. The timing improvement is not quite as good as originally reported but still decent: ~ 1.53 s / 1.27 s = 1.2x

```
$ h5diff -r test_base.h5 test_thisbranch.h5
attribute: <created of </_header>> and <created of </_header>>
size:           H5S_SCALAR           H5S_SCALAR
position        created of </_header> created of </_header> difference          
------------------------------------------------------------
                1.61481e+09     1.61481e+09     326.928        
1 differences found
attribute: <modified of </_header>> and <modified of </_header>>
size:           H5S_SCALAR           H5S_SCALAR
position        modified of </_header> modified of </_header> difference          
------------------------------------------------------------
                1.61481e+09     1.61481e+09     99.9335        
1 differences found
attribute: <version of </_header>> and <version of </_header>>
0 differences found
dataset: </mc_packets_assn> and </mc_packets_assn>
0 differences found
dataset: </messages> and </messages>
0 differences found
dataset: </packets> and </packets>
0 differences found
attribute: <packet_types of </packets>> and <packet_types of </packets>>
0 differences found
--------------------------------
Some objects are not comparable
--------------------------------
Use -c for a list of objects.
```

This branch:

```
$ git checkout dmargala-day3-track-pixel-map
$ python cli/simulate_pixels.py --input_filename examples/edepsim_1M.h5 --output_filename=test_thisbranch.h5 --n_tracks=2000 --pixel_layout=larndsim/pixel_layouts/layout-singlecube.yaml --detector_properties=larndsim/detector_properties/singlecube.yaml
...
*******************
STARTING SIMULATION
*******************
Quenching electrons... 0.52 s
Drifting electrons... 0.41 s
Simulating pixels...: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:07<00:00,  1.30it/s]
- total time: 7.67 s
- excluding first iteration: 1.27 s
Writing to HDF5...: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20096/20096 [00:00<00:00, 22452.23it/s]
Output saved in: test_thisbranch.h5
run_simulation elapsed time: 13.78 s
```

Base:

```
$ git checkout dmargala-day3-add-timing
$ python cli/simulate_pixels.py --input_filename examples/edepsim_1M.h5 --output_filename=test_base.h5 --n_tracks=2000 --pixel_layout=larndsim/pixel_layouts/layout-singlecube.yaml --detector_properties=larndsim/detector_properties/singlecube.yaml
...
*******************
STARTING SIMULATION
*******************
Quenching electrons... 0.63 s
Drifting electrons... 0.56 s
Simulating pixels...: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:08<00:00,  1.14it/s]
- total time: 8.74 s
- excluding first iteration: 1.53 s
Writing to HDF5...: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20096/20096 [00:00<00:00, 22444.31it/s]
Output saved in: test_base.h5
run_simulation elapsed time: 16.38 s
```